+++ apps/MAApp/MAApp.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace MAApp;
+
+use iumioFramework\Core\Requirement\AppModel as App;
+
+/**
+ * Class MAApp
+ * @package MAApp
+ */
+
+class MAApp extends App
+{
+    /*** REGISTER TO APP ***/
+}