+++ elements/compiled/dev/smarty_compiled/a4b5ca6d0f16492b5758e12dc35e02e3583c9161_0.file.modal.tpl.php
@@ -1,11 +1,11 @@
 <?php
-/* Smarty version 3.1.31, created on 2017-12-27 15:30:17
+/* Smarty version 3.1.31, created on 2017-12-29 09:11:57
   from "/Applications/MAMP/htdocs/iumio-framework/vendor/iumio_framework/php/BaseApps/ManagerApp/Front/views/partials/modal.tpl" */
 
 /* @var Smarty_Internal_Template $_smarty_tpl */
 if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   'version' => '3.1.31',
-  'unifunc' => 'content_5a43ae79462d07_85603233',
+  'unifunc' => 'content_5a45f8cdbb74b0_81129049',
   'has_nocache_code' => false,
   'file_dependency' => 
   array (
@@ -20,7 +20,7 @@ if ($_smarty_tpl->_decodeProperties($_smarty_tpl, array (
   array (
   ),
 ),false)) {
-function content_5a43ae79462d07_85603233 (Smarty_Internal_Template $_smarty_tpl) {
+function content_5a45f8cdbb74b0_81129049 (Smarty_Internal_Template $_smarty_tpl) {
 ?>
 <!-- Modal -->
 <div class="modal fade" id="modalManager" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">