+++ web/Dev.php
@@ -4,7 +4,7 @@
  *
  * This is an iumio Framework component
  *
- * (c) RAFINA DANY <danyrafina@gmail.com>
+ * (c) RAFINA DANY <dany.rafina@iumio.com>
  *
  * iumio Framework - iumio Components
  *
@@ -32,7 +32,7 @@ use iumioFramework\Core\Additionnal\TaskBar\TaskBar as TB;
  * @category Framework
  * @licence  MIT License
  * @link https://framework.iumio.com
- * @author   RAFINA Dany <danyrafina@gmail.com>
+ * @author   RAFINA Dany <dany.rafina@iumio.com>
  */
 
 class Dev extends FrameworkEnvironment